requirements.txt
Django==2.0.13
django-crispy-forms==1.7.2
django-extensions==2.1.9
mysqlclient==1.4.2.post1
Pillow==6.0.0
pytz==2019.1
six==1.12.0
